Move utilities for rendering questions for AI grading
